fix review-comments
