+++ src/main/java/com/github/seratch/jslack/Slack.java
@@ -5,8 +5,8 @@ import com.github.seratch.jslack.api.methods.SlackApiException;
 import com.github.seratch.jslack.api.methods.impl.MethodsClientImpl;
 import com.github.seratch.jslack.api.methods.request.rtm.RTMStartRequest;
 import com.github.seratch.jslack.api.rtm.RTMClient;
-import com.github.seratch.jslack.api.scim.ScimClient;
-import com.github.seratch.jslack.api.scim.ScimClientImpl;
+import com.github.seratch.jslack.api.scim.SCIMClient;
+import com.github.seratch.jslack.api.scim.SCIMClientImpl;
 import com.github.seratch.jslack.api.webhook.Payload;
 import com.github.seratch.jslack.api.webhook.WebhookResponse;
 import com.github.seratch.jslack.common.http.SlackHttpClient;
@@ -76,8 +76,8 @@ public class Slack {
     /**
      * Creates a SCIM api client.
      */
-    public ScimClient scim() {
-        return new ScimClientImpl(httpClient);
+    public SCIMClient scim() {
+        return new SCIMClientImpl(httpClient);
     }
 
     /**
