--- src/main/java/com/github/seratch/jslack/api/scim/ScimClientImpl.java
@@ -1,36 +1,0 @@
-package com.github.seratch.jslack.api.scim;
-
-import com.github.seratch.jslack.api.methods.SlackApiException;
-import com.github.seratch.jslack.api.scim.request.UsersDeleteRequest;
-import com.github.seratch.jslack.api.scim.response.UsersDeleteResponse;
-import com.github.seratch.jslack.common.http.SlackHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
-
-import java.io.IOException;
-
-public class ScimClientImpl implements ScimClient {
-  private String endpointUrlPrefix = "https://api.slack.com/scim/v1/Users";
-
-  private final SlackHttpClient slackHttpClient;
-
-  public ScimClientImpl(SlackHttpClient slackHttpClient) {
-    this.slackHttpClient = slackHttpClient;
-  }
-
-  @Override
-  public void setEndpointUrlPrefix(String endpointUrlPrefix) {
-    this.endpointUrlPrefix = endpointUrlPrefix;
-  }
-
-  public UsersDeleteResponse delete(UsersDeleteRequest req) throws IOException, SlackApiException {
-    Request.Builder request = new Request.Builder().url(endpointUrlPrefix + "/" + req.getId()).addHeader("Authorization", "Bearer " + req.getToken());
-    return doPostForm(request, UsersDeleteResponse.class);
-  }
-
-  private <T> T doPostForm(Request.Builder requestBuilder, Class<T> clazz) throws IOException, SlackApiException {
-    Response response = slackHttpClient.delete(requestBuilder.method("DELETE", null).build());
-    return SlackHttpClient.buildJsonResponse(response, clazz);
-  }
-
-}
