+++ src/main/java/com/github/seratch/jslack/api/scim/response/UsersDeleteResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class UsersDeleteResponse implements SlackApiResponse {
-  private boolean ok;
-  private String warning;
-  private String error;
+    private boolean ok;
+    private String warning;
+    private String error;
 }
